move create city to /plans/city
