+++ public/swagger.yaml
@@ -221,6 +221,48 @@ paths:
                 }
           headers: {}
       deprecated: false
+    post:
+      tags:
+        - Account
+      summary: Create new City
+      operationId: CreateCity
+      parameters: []
+      requestBody:
+        description: ''
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/AccountCityRequestBody'
+            example: |
+              {
+                "placeId": "place-id-sdfsd",
+                "name": "Oakland"
+              }
+        required: true
+      responses:
+        '200':
+          description: Successful response
+          content:
+            application/json:
+              example: |
+                {
+                  "_id": "city-id-asdfsd",
+                  "placeId": "place-id-sdfsd",
+                  "name": "Oakland",
+                  "state": "California",
+                  "country": "United States",
+                  "imageURL": "http://example.com/oakland_city.jpg",
+                  "location": [37.8043, -122.2711],
+                  "viewport": {
+                    "low": [37.6317, -122.3558],
+                    "high": [37.8853, -122.1147]
+                  },
+                  "plans": 1,
+                  "createdAt": "2025-10-15T20:18:50.502Z",
+                  "updatedAt": "2025-10-15T20:28:29.754Z"
+                }
+          headers: {}
+      deprecated: false
   /plans/city/{cityId}:
     parameters:
       - in: path
@@ -956,51 +998,6 @@ paths:
       responses:
         '204':
           description: 'No Content'
-          headers: {}
-      deprecated: false
-  /account/cities:
-    post:
-      tags:
-        - Account
-      summary: Create new City
-      security:
-        - bearerAuth: []
-      operationId: CreateCity
-      parameters: []
-      requestBody:
-        description: ''
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/AccountCityRequestBody'
-            example: |
-              {
-                "placeId": "place-id-sdfsd",
-                "name": "Oakland"
-              }
-        required: true
-      responses:
-        '200':
-          description: Successful response
-          content:
-            application/json:
-              example: |
-                {
-                  "_id": "city-id-asdfsd",
-                  "placeId": "place-id-sdfsd",
-                  "name": "Oakland",
-                  "state": "California",
-                  "country": "United States",
-                  "imageURL": "http://example.com/oakland_city.jpg",
-                  "location": [37.8043, -122.2711],
-                  "viewport": {
-                    "low": [37.6317, -122.3558],
-                    "high": [37.8853, -122.1147]
-                  },
-                  "plans": 1,
-                  "createdAt": "2025-10-15T20:18:50.502Z",
-                  "updatedAt": "2025-10-15T20:28:29.754Z"
-                }
           headers: {}
       deprecated: false
 components:
