+++ src/controllers/cities.ts
@@ -8,9 +8,45 @@ import NotFoundError from '../errors/not_found'
 import { geoJsonToCoords } from '../utils/location'
 import { fetchCityDetail } from '../utils/googlePlace'
 import { uploadImage } from '../utils/cloudStorage'
+import CustomAPIError from '../errors/custom_error'
 
 const PAGE_SIZE = 10
 const SORT = '-plans'
+
+type CityDTO = Pick<ICity, 'placeId' | 'name'>
+
+const createNewCity = async (req: Request, res: Response) => {
+  let city: CityDTO = req.body
+  let cityDoc: ICity | null = await CitySchema.findOne({ placeId: city.placeId }).lean()
+
+  // Fetch city details if city does not exist or missing some details
+  if (!cityDoc || (cityDoc && !cityDoc.country && !cityDoc.location)) {
+    let cityDetail = await fetchCityDetail(city.placeId)
+
+    // Store the image
+    cityDetail.imageURL = cityDoc?.imageURL || (await uploadImage(cityDetail.imageURL)) || ''
+
+    // Then update the city document in Mongodb with new data
+    // if doc with placeId exist update, if not create doc
+    await CitySchema.updateOne(
+      { placeId: city.placeId },
+      {
+        $set: {
+          ...cityDetail,
+          location: coordsToGeoJson(cityDetail.location),
+        },
+      },
+      { upsert: true }
+    ).orFail(new CustomAPIError('Failed to create or update the city', StatusCodes.INTERNAL_SERVER_ERROR))
+
+    cityDoc = await CitySchema.findOne({ placeId: city.placeId }).lean()
+  }
+
+  res.status(StatusCodes.CREATED).json({
+    ...cityDoc,
+    location: geoJsonToCoords(cityDoc?.location),
+  })
+}
 
 const fetchAllCities = async (req: Request, res: Response) => {
   const { search, cityId, page = '1', size = PAGE_SIZE, sort = SORT } = req.query
@@ -139,4 +175,4 @@ const attachBookmarkFlagToPlans = async (plans: IPlan[], userId: string) => {
   }))
 }
 
-export { fetchAllCities, fetchCityWithPlans }
+export { createNewCity, fetchAllCities, fetchCityWithPlans }
