+++ src/routes/plans.ts
@@ -1,12 +1,12 @@
 import express from 'express'
 import { fetchAllPlans, fetchPlan, fetchUserWithPlans, fetchAllNearbyPlans } from '../controllers/plans'
-import { fetchCityWithPlans, fetchAllCities } from '../controllers/cities'
+import { fetchCityWithPlans, fetchAllCities, createNewCity } from '../controllers/cities'
 
 const router = express.Router()
 
 router.route('/').get(fetchAllPlans)
 // Specific routes first
-router.route('/city').get(fetchAllCities)
+router.route('/city').get(fetchAllCities).post(createNewCity)
 router.route('/city/:cityId').get(fetchCityWithPlans)
 router.route('/user/:userId').get(fetchUserWithPlans)
